Refactor battle playback and item data handling
- Updated BattlePlayback to clear skill data for heroes regardless of character status.
- Changed health cost calculation in BattlePlayback to use skill's health cost instead of damage.
- Added healthCost, manaCost, and spiritCost properties to ItemDataDto for better resource management.
- Enhanced DataMapper to include healthCost, manaCost, and spiritCost when mapping skill data.
- Fixed potential division issues in ViralUI by ensuring float division for slider values.

+++ Assets/_Project/Scripts/UI/InWorlds/ViralUI.cs
@@ -19,15 +19,15 @@ public class ViralUI : TGTHMonoBehaviour
         switch (type)
         {
             case VitalType.Health:
-                m_viralHealthSlider.value = curValue / maxValue;
+                m_viralHealthSlider.value = (float)curValue / maxValue;
                 m_viralHealthText.text = curValue.ToString();
                 break;
             case VitalType.Mana:
-                m_viralManaSlider.value = curValue / maxValue;
+                m_viralManaSlider.value = (float)curValue / maxValue;
                 m_viralManaText.text = curValue.ToString();
                 break;
             case VitalType.Spirit:
-                m_viralSpiritSlider.value = curValue / maxValue;
+                m_viralSpiritSlider.value = (float)curValue / maxValue;
                 m_viralSpiritText.text = curValue.ToString();
                 break;
         }
